Entry List refactored
