+++ app/src/main/java/journal/gratitude/com/gratitudejournal/ui/timeline/TimelineViewModel.kt
@@ -4,15 +4,14 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import journal.gratitude.com.gratitudejournal.model.Entry
 import journal.gratitude.com.gratitudejournal.model.Milestone
 import journal.gratitude.com.gratitudejournal.model.Milestone.Companion.isMilestone
 import journal.gratitude.com.gratitudejournal.model.TimelineItem
 import journal.gratitude.com.gratitudejournal.repository.EntryRepository
+import journal.gratitude.com.gratitudejournal.util.appendTodayAndYesterday
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import org.threeten.bp.LocalDate
@@ -33,66 +32,25 @@ class TimelineViewModel @Inject constructor(private val repository: EntryReposit
     init {
         scope.launch {
             repository.getEntriesFlow().collect { list ->
-                val today = LocalDate.now()
-                val yesterday = LocalDate.now().minusDays(1)
-                when {
-                    list.isEmpty() -> {
-                        entries.value = listOf<TimelineItem>(
-                            Entry(today, ""),
-                            Entry(yesterday, "")
-                        )
-                    }
-                    list.size < 2 -> {
-                        //user has only ever written one day
-                        val newList = mutableListOf<TimelineItem>()
-                        newList.addAll(list)
-                        if (list[0].entryDate != today) {
-                            newList.add(0, Entry(today, ""))
+                val listWithAppendedTodayAndYesterday = appendTodayAndYesterday(list)
+                val listWithHintsAndMilestones = mutableListOf<TimelineItem>()
+                var numEntries = 0
+                for (index in listWithAppendedTodayAndYesterday.size - 1 downTo 0) {
+                    listWithHintsAndMilestones.add(0, listWithAppendedTodayAndYesterday[index])
+                    if (listWithAppendedTodayAndYesterday[index].entryContent.isNotEmpty()) {
+                        numEntries++
+                        if (isMilestone(numEntries)) {
+                            listWithHintsAndMilestones.add(0, Milestone.create(numEntries))
                         }
-                        if (list[0].entryDate != yesterday) {
-                            newList.add(1, Entry(yesterday, ""))
-                        }
-                        entries.value = newList
-                    }
-                    else -> {
-                        val latest = list[0]
-                        val listWithHints = mutableListOf<Entry>()
-                        listWithHints.addAll(list)
-                        if (latest.entryDate != today) {
-                            //they dont have the latest
-                            listWithHints.add(0, Entry(today, ""))
-                        }
-                        if (listWithHints[1].entryDate != yesterday) {
-                            listWithHints.add(1, Entry(yesterday, ""))
-                        }
-
-                        val listWithHintsAndMilestones = mutableListOf<TimelineItem>()
-
-                        var numEntries = 0
-                        for (index in listWithHints.size - 1 downTo 0) {
-                            listWithHintsAndMilestones.add(0, listWithHints[index])
-                            if (listWithHints[index].entryContent.isNotEmpty()) {
-                                numEntries++
-                                if (isMilestone(numEntries)) {
-                                    listWithHintsAndMilestones.add(0, Milestone.create(numEntries))
-                                }
-                            }
-
-                        }
-                        entries.value = listWithHintsAndMilestones
                     }
                 }
+                entries.value = listWithHintsAndMilestones
             }
-
         }
     }
 
     fun getTimelineItems(): List<TimelineItem> {
         return entries.value ?: emptyList()
-    }
-
-    fun addEntries(entries: List<Entry>) = scope.launch(Dispatchers.IO) {
-        repository.addEntries(entries)
     }
 
     override fun onCleared() {
